refactor(fe): extract PhaseSteps and TechnicalDetails
