+++ frontend/src/components/AdoptNodeModal.tsx
@@ -2,7 +2,6 @@ import {
   AlertCircle,
   Camera,
   CheckCircle2,
-  Circle,
   ClipboardPaste,
   Loader2,
   RotateCw,
@@ -16,7 +15,7 @@ import { useGateways } from "../hooks/useGateways";
 import { ApiError, apiJson } from "../lib/api";
 import { cn } from "../lib/utils";
 import { humanizeAdoptError, mapAdoptError } from "./adopt/lib/errors";
-import { describePhase, humanizePhase, PHASE_ORDER } from "./adopt/lib/phases";
+import { humanizePhase } from "./adopt/lib/phases";
 import { extractQrPayload, validateQrPayload } from "./adopt/lib/qr";
 import { ErrorBlock } from "./adopt/shared/ErrorBlock";
 import { ModeTab } from "./adopt/shared/ModeTab";
@@ -24,6 +23,8 @@ import { SectionHead } from "./adopt/shared/SectionHead";
 import { Skeleton } from "./adopt/shared/Skeleton";
 import { BerthStep } from "./adopt/steps/BerthStep";
 import { CameraScan } from "./adopt/steps/CameraScan";
+import { PhaseSteps } from "./adopt/steps/PhaseSteps";
+import { TechnicalDetails } from "./adopt/steps/TechnicalDetails";
 
 type Berth = components["schemas"]["BerthOut"];
 type Gateway = components["schemas"]["GatewayOut"];
@@ -554,176 +555,5 @@ function ProgressStep({
         </button>
       )}
     </div>
-  );
-}
-
-// row pitch in px, must match the inline `height` on each <li> below
-const PHASE_ROW = 26;
-// rows kept visible above the focused one (sets the "scrolled" feel)
-const PHASE_HEAD_ROOM = 1;
-// total visible rows in the window (focused + head room + 1 lookahead)
-const PHASE_VISIBLE_ROWS = PHASE_HEAD_ROOM + 2;
-
-function PhaseSteps({
-  phase,
-  status,
-}: {
-  phase: string | null;
-  status: "pending" | "ok" | "err";
-}) {
-  const currentIdx = phase ? PHASE_ORDER.indexOf(phase as never) : -1;
-  // peg window to the last phase once finalized so the "complete" row settles
-  // at the focused position instead of the list snapping back
-  const focusIdx =
-    status === "ok" ? PHASE_ORDER.length - 1 : Math.max(currentIdx, 0);
-  const offsetPx = Math.max(0, focusIdx - PHASE_HEAD_ROOM) * PHASE_ROW;
-
-  return (
-    <div
-      className="relative overflow-hidden"
-      style={{
-        height: PHASE_ROW * PHASE_VISIBLE_ROWS,
-        // older rows fade out at the top so the list reads as scrolling
-        maskImage:
-          "linear-gradient(to bottom, transparent 0, black 28%, black 100%)",
-        WebkitMaskImage:
-          "linear-gradient(to bottom, transparent 0, black 28%, black 100%)",
-      }}
-    >
-      {/* timeline rail behind the icons, fades with the row mask above */}
-      <div
-        aria-hidden
-        className="absolute left-[11px] top-0 bottom-0 w-px bg-brand-navy/10"
-      />
-      <ol
-        aria-label="provisioning phases"
-        className="flex flex-col transition-transform duration-500 ease-out"
-        style={{ transform: `translateY(-${offsetPx}px)` }}
-      >
-        {PHASE_ORDER.map((p, i) => {
-          const done = status === "ok" || i < currentIdx;
-          const active = status === "pending" && i === currentIdx;
-          const failed = status === "err" && i === currentIdx;
-          // distance from the focused row drives the secondary fade so rows
-          // ahead/behind dim further the more they recede
-          const distance = Math.abs(i - focusIdx);
-          const distanceOpacity =
-            distance === 0 ? 1 : distance === 1 ? 0.55 : 0.2;
-          const detail = describePhase(p);
-          return (
-            <li
-              key={p}
-              aria-current={active ? "step" : undefined}
-              title={detail}
-              className="group relative flex items-center gap-3 transition-opacity duration-500 ease-out pl-1 cursor-help"
-              style={{ height: PHASE_ROW, opacity: distanceOpacity }}
-            >
-              <span
-                className={cn(
-                  "relative z-10 grid place-items-center w-5 h-5 rounded-full bg-white",
-                  "transition-transform duration-300 ease-out",
-                  active && "scale-110",
-                )}
-              >
-                {done ? (
-                  <CheckCircle2
-                    size={16}
-                    strokeWidth={2.5}
-                    className="text-emerald-600"
-                  />
-                ) : active ? (
-                  <Loader2
-                    size={16}
-                    strokeWidth={2.5}
-                    className="text-brand-blue animate-spin"
-                  />
-                ) : failed ? (
-                  <AlertCircle
-                    size={16}
-                    strokeWidth={2.5}
-                    className="text-red-600"
-                  />
-                ) : (
-                  <Circle
-                    size={14}
-                    strokeWidth={2.5}
-                    className="text-brand-navy/40"
-                  />
-                )}
-              </span>
-              <span
-                className={cn(
-                  "text-[10px] font-bold uppercase tracking-widest transition-colors",
-                  done && "text-emerald-700",
-                  active && "text-brand-blue text-[11px]",
-                  failed && "text-red-700",
-                  !done && !active && !failed && "text-brand-navy/60",
-                )}
-              >
-                {humanizePhase(p)}
-              </span>
-              {/* hover detail, positioned over the row so it never reflows.
-                  z-30 sits above the mask gradient so older rows can still
-                  reveal their detail on hover */}
-              <span
-                role="tooltip"
-                className={cn(
-                  "pointer-events-none absolute left-9 top-1/2 -translate-y-1/2 z-30",
-                  "px-2.5 py-1 rounded-md bg-brand-navy text-white",
-                  "text-[10px] normal-case tracking-normal font-medium leading-tight",
-                  "max-w-[260px] whitespace-normal shadow-lg",
-                  "opacity-0 -translate-x-1 transition-all duration-150 delay-100 ease-out",
-                  "group-hover:opacity-100 group-hover:translate-x-0 group-focus-visible:opacity-100",
-                )}
-              >
-                {detail}
-              </span>
-            </li>
-          );
-        })}
-      </ol>
-    </div>
-  );
-}
-
-function TechnicalDetails({
-  requestId,
-  request,
-  gateway,
-  berth,
-}: {
-  requestId: string;
-  request: components["schemas"]["AdoptionRequestOut"] | null;
-  gateway: Gateway | null;
-  berth: Berth | null;
-}) {
-  const rows: Array<[string, string | null | undefined]> = [
-    ["Request ID", requestId],
-    ["Mesh UUID", request?.mesh_uuid],
-    ["Mesh unicast addr", request?.mesh_unicast_addr],
-    ["Berth", berth?.berth_id ?? request?.berth_id],
-    ["Gateway", gateway?.gateway_id ?? request?.gateway_id],
-    ["Serial number", request?.serial_number],
-  ].filter(([, v]) => Boolean(v));
-  if (rows.length === 0) return null;
-  return (
-    <details className="group rounded-xl border border-brand-navy/10 bg-brand-navy/[0.02] open:bg-brand-navy/[0.04]">
-      <summary className="cursor-pointer list-none px-4 py-2 text-[10px] font-bold uppercase tracking-widest text-brand-navy/50 hover:text-brand-navy/80 select-none flex items-center justify-between">
-        <span>Technical details</span>
-        <span className="text-brand-navy/30 group-open:rotate-180 transition-transform">
-          ▾
-        </span>
-      </summary>
-      <dl className="px-4 pb-3 grid grid-cols-[auto_1fr] gap-x-4 gap-y-1 text-[10px]">
-        {rows.map(([label, value]) => (
-          <div key={label} className="contents">
-            <dt className="text-brand-navy/50 font-bold uppercase tracking-widest whitespace-nowrap">
-              {label}
-            </dt>
-            <dd className="text-brand-navy/80 font-mono break-all">{value}</dd>
-          </div>
-        ))}
-      </dl>
-    </details>
   );
 }
